+++ src/Libraries/SmartStore.Data/Migrations/MigrationsConfiguration.cs
@@ -918,7 +918,7 @@
             builder.AddOrUpdate("Admin.Configuration.Settings.GeneralCommon.SocialSettings.FacebookAppId",
                 "Facebook App ID",
                 "Facebook App ID",
-                "A Facebook App ID is a unique number that identifies your application and links it to your Facebook account. It is used when sharing products, category and manufacturer pages on Facebook.",
+                "A Facebook App ID is a unique number that identifies your application and connects it to your Facebook account. It is used when sharing products, category and manufacturer pages on Facebook.",
                 "Eine Facebook App ID ist eine eindeutige Nummer, die Ihre Anwendung identifiziert und mit Ihrem Facebook-Account verknüpft. Wird beim Teilen von Produkten, Warengruppen- und Herstellerseiten auf Facebook genutzt.");
 
             builder.AddOrUpdate("Admin.Configuration.Settings.GeneralCommon.SocialSettings.TwitterSite",